requirements.txt
fastapi==0.115.6
uvicorn[standard]==0.34.0
mistralai==1.5.0
python-dotenv==1.0.1
pydantic==2.10.5
